// citpwawcl/src/app/models/controllo-disponibile.ts
export class ControlloDisponibileModel {
    nomeComponente: string;
    dataControllo: string;
    tipoControllo: string;
    tipoComponente: string;
    dataInstall: string;
    descMarca: string;
    modello: string;
    matricola: string;
    descDettaglio: string;
    potTermica: string;
    descCombustibile: string;
    potTermicaRaff: number;
    potTermicaRisc: number;
    alimentazione: string;
    coMin: number;
    coMax: number;
    progressivo: number;
    nModuli: number;
    cfPIvaImpresa: string;
    siglaReaImpresa: string;
    numeroReaImpresa: string;
}