// citpwawcl/src/app/models/manut-form-model.ts
export class ManutFormModel {
    dataControllo: string;
    tipoControllo: string;
    tipoComponente: string;
    tipoIntervento: string;
    osservazioni: string;
    raccomandazioni: string;
    prescrizioni: string;
    prossimoInterventoEntro: string;
    nomeCognomeTecnico: string;
    oraArrivo: string;
    oraPartenza: string;
    progressivo: number;
    codiceFiscale: string;
    numeroRea: string;
    siglaRea: string;
}